fix long

+++ src/fr/eni/clinique/dal/AnimauxDAOJdbcImpl.java
@@ -49,7 +49,7 @@ public class AnimauxDAOJdbcImpl implements DaoAnimaux{
 	        stmt.setString(3,a1.getCouleur());
 	        stmt.setString(4,a1.getRace());
 	        stmt.setString(5,a1.getEspece());
-	        stmt.setInt(6, a1.getCodeClient());
+	        stmt.setLong(6, a1.getCodeClient());
 	        stmt.setString(7,a1.getTatouage());
 	        stmt.setString(8,a1.getAntecedents());
 	        stmt.setBoolean(9, a1.isArchive());
@@ -193,7 +193,7 @@ public class AnimauxDAOJdbcImpl implements DaoAnimaux{
 			stmt.setString(3,a1.getCouleur());
 			stmt.setString(4,a1.getRace());
 			stmt.setString(5,a1.getEspece());
-			stmt.setInt(6,a1.getCodeClient());
+			stmt.setLong(6,a1.getCodeClient());
 			stmt.setString(7,a1.getTatouage());
 			stmt.setString(8,a1.getAntecedents());
 			stmt.setBoolean(9, a1.isArchive());
